Uploading files, tests

+++ src/Sher.Application/Files/DeleteFile/DeleteFileCommandHandler.cs
@@ -1,11 +1,11 @@
 using System.Threading;
 using System.Threading.Tasks;
-using MediatR;
+using Sher.Application.Processing;
 using Sher.Core.Files;
 
 namespace Sher.Application.Files.DeleteFile
 {
-    public class DeleteFileCommandHandler : AsyncRequestHandler<DeleteFileCommand>
+    public class DeleteFileCommandHandler : ICommandHandler<DeleteFileCommand, bool>
     {
         private readonly IFileRepository _fileRepository;
         private readonly IUploaderRepository _uploaderRepository;
@@ -16,13 +16,19 @@ namespace Sher.Application.Files.DeleteFile
             _uploaderRepository = uploaderRepository;
         }
 
-        protected override async Task Handle(DeleteFileCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteFileCommand request, CancellationToken cancellationToken)
         {
             var (fileId, uploaderId) = request;
             var uploader = await _uploaderRepository.GetByIdAsync(uploaderId);
             var file = await _fileRepository.GetByIdAsync(fileId);
 
+            if (uploader is null || file is null)
+            {
+                return false;
+            }
+
             uploader.DeleteFile(file);
+            return true;
         }
     }
 }
